rembg-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
rembg==2.0.50
pillow==10.1.0
pillow-heif==0.14.0
gunicorn==21.2.0

